block_svg_framed
